Fix ActivePrice and ExpiringTpr bugs, and log messages.
1. The log messages had an unnecessary `$` at the start.
1. Bug fixes
   1. Fixed schema data types and as a result the DB models.
   1. The last subset of ActivePrice and ExpiringTprs were not processed.
   1. JustInTime archive was using wrong parameter.
   1. DB transaction was not committed.
   1. MammothPrices object didn't have the array of MammothPrice assigned.

Related work items: #51280

+++ Icon/GPM/GPMService/GPMService.Producer/Model/DBModel/GetActivePricesQueryModel.cs
@@ -1,11 +1,10 @@
 ﻿using System;
-using System.Numerics;
 
 namespace GPMService.Producer.Model.DBModel
 {
     internal class GetActivePricesQueryModel
     {
-        public BigInteger PriceID { get; set; }
+        public long PriceID { get; set; }
         public string Region { get; set; }
         public string GpmID { get; set; }
         public int ItemID { get; set; }
@@ -17,7 +16,7 @@ namespace GPMService.Producer.Model.DBModel
         public string PriceTypeAttribute { get; set; }
         public string SellableUOM { get; set; }
         public string CurrencyCode { get; set; }
-        public int Multiple { get; set; }
+        public byte Multiple { get; set; }
         public DateTime? TagExpirationDate { get; set; }
         public DateTime InsertDateUtc { get; set; }
         public DateTime? ModifiedDateUtc { get; set; }
